feature: link `composer.lock` to working directory
When running psalm through this codeception module, the working directory changes. In the working directory psalm is operating on, the projects `composer.lock` is unavailable and thus, psalm won't cache.
When linking the projects `composer.lock` to the working directory, psalm starts using its cache which have a positive impact on build times in combination with this module.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>